CID-1864 Enhance Dev Mode

+++ src/main/kotlin/net/leanix/vsm/sbomBooster/service/ProcessService.kt
@@ -83,8 +83,11 @@ class ProcessService(
             } catch (e: Exception) {
                 logger.error(e.message)
             } finally {
-                ortService.deleteDownloadedFolder(downloadedFolder)
-                logger.info("Finished deleting temp folder $downloadedFolder.")
+                if (!propertiesConfiguration.devMode) {
+                    logger.info("Beginning to delete folder $downloadedFolder.")
+                    ortService.deleteDownloadedFolder(downloadedFolder)
+                    logger.info("Finished deleting temp folder $downloadedFolder.")
+                }
             }
         }
